Refactored get_correctional_factors function into the GetCorrectionalFactors class.
